current temperature data

+++ WeatherCube/Assets/Scripts/Response/Response.Forecast.cs
@@ -14,7 +14,7 @@ namespace Response.Forecast {
 
 	public class Main
 	{
-		public double temp { get; set; }
+		public float temp { get; set; }
 		public double feels_like { get; set; }
 		public double temp_min { get; set; }
 		public double temp_max { get; set; }
